Add auotscreen tracking functionality
