+++ example-app/src/main/AndroidManifest.xml
@@ -11,6 +11,8 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
+        <meta-data android:name="rvAutoTrackingEnabled" android:value="true"></meta-data>
+
         <service android:name="io.rover.example.fcm.FirebaseMessageReceiver">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
@@ -25,6 +27,9 @@
         <activity
             android:name="io.rover.example.ExampleMainActivity"
             android:label="@string/app_name" android:theme="@style/RoverDebugPreferencesActivityTheme">
+
+            <meta-data android:name="rvAutoTrackingExcludeActivity" android:value="false"></meta-data>
+            <meta-data android:name="rvAutoTrackingLabel" android:value="Main Activity"></meta-data>
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
